refactor(agents): own persisted budget vocabulary
Budget projection records belong with the agent state they describe,
while the harness retains parsing and budget policy.
